pyproject.toml
[project]
name = "plotforge"
version = "0.2.1"
description = "PlotForge - Web-based SVG to G-code converter"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "vpype>=1.14.0",
    "vpype-gcode>=0.4.0",
    "pillow>=10.0",
    "opencv-python-headless>=4.9",
    "numpy>=1.26",
    "vtracer>=0.6.11",
    "rembg>=2.0.56",
    "onnxruntime>=1.17",
]

